+++ src/api/caidapp/urls.py
@@ -360,6 +360,8 @@ urlpatterns = [
     # path("generic/locality/", ListView.as_view(model=models.Locality), name="generic_localities"),
     path("generic/locality/", views.LocalityListView.as_view(), name="generic_locality_list"),
     path("generic/locality/<int:pk>/", DetailView.as_view(model=models.Locality), name="generic_locality_detail"),
+    path("suggest_merge_localities/", views_locality.suggest_merge_localities_view, name="suggest_merge_localities"),
+    path("merge_localities/<int:locality_from_id>/<int:locality_to_id>/", views_locality.merge_localities_view, name="merge_localities"),
 
 ]
 # if settings.DEBUG: